spectral and mux code -- written as of right now to just test an individual spectral sensor
